refactor: update project components imports to use shadcn directory
